Fixed segfault RSPreviewWidget.

+++ src/rs-preview-widget.c
@@ -432,7 +432,7 @@ rs_preview_widget_set_photo(RSPreviewWidget *preview, RS_PHOTO *photo)
 			g_object_set(preview->filter_input[view], "image", preview->photo->input, NULL);
 			g_object_set(preview->filter_rotate[view], "angle", preview->photo->angle, "orientation", preview->photo->orientation, NULL);
 			g_object_set(preview->filter_crop[view], "rectangle", preview->photo->crop, NULL);
-			g_object_set(preview->filter_denoise[view], "sharpen", preview->scale * preview->photo->settings[preview->snapshot[view]]->sharpen, NULL);
+			g_object_set(preview->filter_denoise[view], "sharpen", (gint) (preview->scale * preview->photo->settings[preview->snapshot[view]]->sharpen), NULL);
 			rescale(preview, view);
 		}
 	}
@@ -669,7 +669,7 @@ rs_preview_widget_set_snapshot(RSPreviewWidget *preview, const guint view, const
 
 	rs_color_transform_set_from_settings(preview->rct[view], preview->photo->settings[preview->snapshot[view]], MASK_ALL);
 
-	g_object_set(preview->filter_denoise[view], "sharpen", preview->scale * preview->photo->settings[preview->snapshot[view]]->sharpen, NULL);
+	g_object_set(preview->filter_denoise[view], "sharpen", (gint) (preview->scale * preview->photo->settings[preview->snapshot[view]]->sharpen), NULL);
 
 	DIRTY(preview->dirty[view], BUFFER);
 	DIRTY(preview->dirty[view], SCREEN);
